Add deploy_site role
The deploy_site role is a base level role that can be used to deploy a
number of sites that run inside of images. One variable `site` is
required. This variable dictates which settings, images, and containers
are pulled from the configuration and used during the deployment phases.

Further documentation was added to the `README.md` on how the role
works.