Add per site environment variable handling
The environments data structure allows for environment variables to be
set based on the site that is being deployed, and via the group_vars, to
which environment (production/test).

This is the first such setting, so subject to change, to set the port
that the github-meets-cpan site listens in for nginx to proxy to.