Add role for updating environment files
The new `env_files` role is to be used when updating a docker
environment file with values that come from Ansible variables/vault.

The role uses two variables `container` and `env` for determining where
and what to set values to. It searches the appropriate `.env` file for
the key value from the Ansible variable, and updates its setting with
the Ansible variable value.