Add output redirection for cron entry
Stop all the emails for docker-compose starting up for the cron job to
populate github-meets-cpan.

Unfortunately, docker-compose has an outstanding pull request
(https://github.com/docker/compose/issues/6026) to fix the functionality
of `--log-level` with the `up` and `run` command so that they can run
silently but still print out errors from the containers. The work around
until such a time is to direct output to `/dev/null` for both STDOUT and
STDERR.