Add system updates playbook
This playbook checks if there are any system level updates that need to
be installed. If there are it installs them, and if they require a
reboot, reboots the server. Checking that it comes back up afterwards
before proceeding.

The task runs serially so that only 1 server at a time would be out.